+++ packages/cubejs-playground/src/QueryBuilder/ButtonDropdown.tsx
@@ -1,8 +1,12 @@
-import { Dropdown } from 'antd';
+import { Dropdown, DropDownProps, ButtonProps } from 'antd';
 
 import { Button } from '../atoms';
 
-const ButtonDropdown = ({ overlay, disabled = false, ...buttonProps }: any) => {
+const ButtonDropdown = ({
+  overlay,
+  disabled = false,
+  ...buttonProps
+}: DropDownProps & ButtonProps) => {
   return (
     <Dropdown
       disabled={disabled}