+++ packages/cubejs-query-orchestrator/src/driver/driver.interface.ts
@@ -54,10 +54,18 @@ export type IndexesSQL = {
   sql: [string, unknown[]];
 }[];
 
+export type UnloadOptions = {
+  maxFileSize: number,
+};
+
+export type QueryOptions = {};
+
 export interface DriverInterface {
   createSchemaIfNotExists(schemaName: string): Promise<any>;
   uploadTableWithIndexes(table: string, columns: TableStructure, tableData: DownloadTableData, indexesSql: IndexesSQL): Promise<void>;
   loadPreAggregationIntoTable: (preAggregationTableName: string, loadSql: string, params: any, options: any) => Promise<any>;
+  //
+  query<R = unknown>(query: string, params: unknown[], options?: QueryOptions): Promise<R[]>;
   //
   tableColumnTypes: (table: string) => Promise<TableStructure>;
   // Download data from Query (for readOnly)
@@ -66,4 +74,8 @@ export interface DriverInterface {
   downloadTable: (table: string, options: ExternalDriverCompatibilities) => Promise<DownloadTableMemoryData | DownloadTableCSVData>;
   // Some drivers can implement streaming from SQL
   stream?: (table: string, values: unknown[], options: StreamOptions) => Promise<StreamTableData>;
+  // Some drivers can implement UNLOAD data to external storage
+  unload?: (table: string, options: UnloadOptions) => Promise<DownloadTableCSVData>;
+  // Some drivers can implement UNLOAD data to external storage
+  isUnloadSupported?: (options: UnloadOptions) => Promise<boolean>;
 }